feat: dashboard muestra resumen de feedback
