+++ client/src/components/CardsFeedback/cardsfeedback.module.css
@@ -1,0 +1,22 @@
+.feedbackComments {
+  width: 500px;
+  border-bottom: 1px solid;
+}
+
+.feedbackAvatar {
+  padding: 0px 10px;
+  display: flex;
+  gap: 10px;
+}
+
+.feedbackComment {
+  padding: 20px 0px;
+  border-top: 1px solid;
+}
+
+.feedbackComment img {
+  width: 50px;
+  height: 50px;
+  background: #fff;
+  border-radius: 50%;
+}
