+++ client/src/components/CardsFeedback/index.jsx
@@ -1,0 +1,2 @@
+export { default as CardsFeedback } from './CardsFeedback';
+export { default as CardsFeedbackCompact } from './CardsFeedbackCompact';
