+++ client/src/pages/Dashboard/index.js
@@ -1,4 +1,5 @@
 import { Link, Route, Routes } from "react-router-dom";
+import { useSelector } from "react-redux";
 
 import s from "./dashboard.module.css";
 import SideBar from "../../components/SideBar";
@@ -6,8 +7,12 @@ import DashboardHome from "./DashboardHome.jsx";
 import DashboardFeedback from "./DashboardFeedback.jsx";
 
 export default function Dashboard() {
+  const { user } = useSelector(state => state.user)
+
   return (
     <div className={s.container}>
+      {Object.keys(user).length ?
+      <>
       <SideBar>
         <Link to=''>Home</Link>
         <br />
@@ -17,6 +22,9 @@ export default function Dashboard() {
         <Route exact path="/" element={<DashboardHome />} />
         <Route exact path="/feedback" element={<DashboardFeedback />} />
       </Routes>
+      </>
+      : 'debe iniciar sesion'
+      }
     </div>
   )
 }
